Image_Captioning/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blobfile==0.11.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
clip==0.2.0
cycler==0.10.0
dataclasses==0.8
decorator==5.0.7
defusedxml==0.7.1
einops==0.3.0
entrypoints==0.3
filelock==3.0.12
fsspec==2021.4.0
ftfy==6.0.1
future==0.18.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
huggingface-hub==0.0.8
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
logging==0.4.9.6
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mpmath==1.2.1
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
ninja==1.10.0.post2
nltk==3.6.2
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.6
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.5.0
sentence-transformers==1.2.0
sentencepiece==0.1.95
six==1.15.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.61.0
traitlets==4.3.3
transformers==4.6.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xmltodict==0.12.0
yarl==1.6.3
zipp==3.4.1
